requirements.txt
PyPDF2~=1.26.0
paramiko~=2.7.2
Pillow~=8.1.0